requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.3.1
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cleanpy==0.5.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
config==0.5.1
cryptography==42.0.8
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.0
django-admin-toolbox==1.0.0.dev29
django-allauth==0.63.3
django-cors-headers==4.4.0
django-countries==7.6.1
django-debug-toolbar==4.4.2
django-jazzmin==3.0.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.15.4
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
gunicorn==22.0.0
httplib2==0.22.0
humanize==4.9.0
identify==2.5.36
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
multidict==6.0.5
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.13.39
phonenumberslite==8.13.39
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
setuptools==70.1.1
six==1.16.0
sqlparse==0.5.0
stripe==10.1.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
